feat: softDelete 적용 및 조건 추가

+++ src/main/java/org/example/baedalteam27/domain/category/service/CategoryService.java
@@ -43,6 +43,7 @@ public class CategoryService {
         List<Category> categories = categoryRepository.findAll();
 
         return categories.stream()
+                .filter(category -> !category.isDeleted())  // isDeleted 가 false인 상태만 조회
                 .map(category -> new FindCategoriesResponseDto(category.getId(), category.getName()))
                 .collect(Collectors.toList());
     }
@@ -62,6 +63,11 @@ public class CategoryService {
             throw new IllegalArgumentException("중복된 이름입니다.");
         }
 
+        // 삭제된 카테고리인지 검증
+        if (category.isDeleted()) {
+            throw new IllegalArgumentException("이미 삭제된 카테고리 입니다.");
+        }
+
         category.update(requestDto.getName());
     }
 
@@ -74,6 +80,11 @@ public class CategoryService {
         }
 
         Category category = categoryRepository.findByIdOrElseThrow(categoryid);
+
+        // 삭제된 카테고리인지 검증
+        if (category.isDeleted()) {
+            throw new IllegalArgumentException("삭제된 카테고리를 다시 삭제할 수 없습니다.");
+        }
 
         categoryRepository.delete(category);
     }
